Change monthlyCost from int to a float for BMPlan

+++ src/Services/Plans/BMPlan.php
@@ -14,7 +14,7 @@ class BMPlan extends Model
 	protected int $disk;
 	protected int $diskCount;
 	protected int $bandwidth;
-	protected int $monthlyCost;
+	protected float $monthlyCost;
 	protected string $type;
 	protected array $locations;
 
@@ -98,12 +98,12 @@ class BMPlan extends Model
 		$this->bandwidth = $bandwidth;
 	}
 
-	public function getMonthlyCost() : int
+	public function getMonthlyCost() : float
 	{
 		return $this->monthlyCost;
 	}
 
-	public function setMonthlyCost(int $monthly_cost) : void
+	public function setMonthlyCost(float $monthly_cost) : void
 	{
 		$this->monthlyCost = $monthly_cost;
 	}
